Refactor out controller
