+++ src/main/java/duke/Main.java
@@ -9,6 +9,7 @@ import duke.command.CommandParser;
 import duke.command.CommandType;
 import duke.commons.exceptions.DukeException;
 import duke.commons.utils.DateUtils;
+import duke.controller.LogicController;
 import duke.storage.PersistentStorageHandler;
 import duke.task.Deadline;
 import duke.task.Event;
@@ -38,9 +39,11 @@ import javafx.stage.Stage;
  */
 public class Main extends Application {
 
-    private static TaskList taskList = new TaskList();
+    private TaskList taskList = new TaskList();
 
-    private static PersistentStorageHandler persistentStorageHandler = new PersistentStorageHandler();
+    private PersistentStorageHandler persistentStorageHandler = new PersistentStorageHandler();
+
+    private LogicController logicController;
 
     private ScrollPane scrollPane;
     private VBox dialogContainer;
@@ -56,6 +59,7 @@ public class Main extends Application {
         initGui(stage);
         initEventHandlers();
         initLoadStorage();
+        initController();
     }
 
     private void initGui(Stage stage) {
@@ -137,7 +141,7 @@ public class Main extends Application {
             String userInput = userInputField.getText();
             createUserDialog(userInput);
 
-            String response = getResponse(userInput);
+            String response = logicController.processUserInput(userInput);
             createAgentDialog(response);
 
             persistentStorageHandler.writeTaskFileToDisc(taskList);
@@ -147,56 +151,6 @@ public class Main extends Application {
         } finally {
             userInputField.clear();
         }
-    }
-
-    private String getResponse(String userInput) throws DukeException {
-        CommandType commandType = CommandParser.parseCommand(userInput);
-        String response = "Invalid Command";
-        try {
-            switch (commandType) {
-                case LIST:
-                    response = handleList();
-                    break;
-
-                case MARK:
-                    response = handleMark(userInput);
-                    break;
-
-                case UNMARK:
-                    response = handleUnmark(userInput);
-                    break;
-
-                case DELETE:
-                    response = handleDelete(userInput);
-                    break;
-
-                case FIND:
-                    response = handleFind(userInput);
-                    break;
-
-                case TODO:
-                    response = handleToDo(userInput);
-                    break;
-
-                case DEADLINE:
-                    response = handleDeadline(userInput);
-                    break;
-
-                case EVENT:
-                    response = handleEvent(userInput);
-                    break;
-
-                case BYE:
-                    response = handleExit();
-                    break;
-
-                default:
-                    throw new DukeException("Invalid Command" + commandType);
-            }
-        } catch (DukeException e) {
-            response = e.getMessage();
-        }
-        return response;
     }
 
     private void createUserDialog(String text) {
@@ -209,116 +163,8 @@ public class Main extends Application {
                 DialogBox.createAgentDialog(new Label(text), new ImageView(derek)));
     }
 
-    /**
-     * Displays the list of tasks to the user.
-     * 
-     * @throws DukeException If an error during task listing.
-     */
-    private static String handleList() throws DukeException {
-        String response = taskList.getFormattedTasks();
-        return UserInterface.formatResponse(response);
-    }
-
-    /**
-     * Marks a specified task as done.
-     * 
-     * @param userInput The user input containing the index of the task to mark as
-     *                  done.
-     * @throws DukeException If the task index is invalid.
-     */
-    private static String handleMark(String userInput) throws DukeException {
-        int idx = CommandParser.parseTaskIndex(userInput);
-        String response = taskList.markTaskDone(idx);
-        return UserInterface.formatResponse(response);
-    }
-
-    /**
-     * Marks a specified task as not done.
-     * 
-     * @param userInput The user input containing the index of the task to mark as
-     *                  not done.
-     * @throws DukeException If the task index is invalid.
-     */
-    private static String handleUnmark(String userInput) throws DukeException {
-        int idx = CommandParser.parseTaskIndex(userInput);
-        String response = taskList.markTaskUndone(idx);
-        return UserInterface.formatResponse(response);
-    }
-
-    /**
-     * Deletes a specified task from the task list.
-     * 
-     * @param userInput The user input containing the index of the task to delete.
-     * @throws DukeException If the task index is invalid.
-     */
-    private static String handleDelete(String userInput) throws DukeException {
-        int idx = CommandParser.parseTaskIndex(userInput);
-        String response = taskList.deleteTask(idx);
-        int totalTasks = taskList.getNumberTasks();
-        return UserInterface.formatTaskDeletedResponse(response, totalTasks);
-    }
-
-    private static String handleFind(String userInput) throws DukeException {
-        String[] keywords = CommandParser.parseFind(userInput);
-        ArrayList<Integer> taskIndices = taskList.findTasksByKeywordsMatching(keywords);
-        ArrayList<String> response = taskList.getTaskRepresentationsByIndices(taskIndices);
-        return UserInterface.formatTasksByIndicesResponse(response);
-    }
-
-    /**
-     * Adds a new ToDo task to the task list.
-     * 
-     * @param userInput The user input containing the description of the ToDo task.
-     * @throws DukeException If the description is invalid or other errors occur.
-     */
-    private static String handleToDo(String userInput) throws DukeException {
-        String description = CommandParser.parseToDo(userInput);
-        String response = taskList.addTask(new ToDo(description));
-        int totalTasks = taskList.getNumberTasks();
-        return UserInterface.formatTaskAddedResponse(response, totalTasks);
-    }
-
-    /**
-     * Adds a new Deadline task to the task list.
-     * 
-     * @param userInput The user input containing the description of the Deadline
-     *                  task.
-     * @throws DukeException If the description is invalid or other errors occur.
-     */
-    private static String handleDeadline(String userInput) throws DukeException {
-        String[] deadlineDetails = CommandParser.parseDeadline(userInput);
-        String description = deadlineDetails[0];
-        LocalDate due = DateUtils.parseDateString(deadlineDetails[1]);
-        String response = taskList.addTask(new Deadline(description, due));
-        int totalTasks = taskList.getNumberTasks();
-        return UserInterface.formatTaskAddedResponse(response, totalTasks);
-    }
-
-    /**
-     * Adds a new Event task to the task list.
-     * 
-     * @param userInput The user input containing the description of the Event task.
-     * @throws DukeException If the description is invalid or other errors occur.
-     */
-    private static String handleEvent(String userInput) throws DukeException {
-        String[] eventDetails = CommandParser.parseEvent(userInput);
-        String description = eventDetails[0];
-        LocalDate start = DateUtils.parseDateString(eventDetails[1]);
-        LocalDate end = DateUtils.parseDateString(eventDetails[2]);
-        String response = taskList.addTask(new Event(description, start, end));
-        int totalTasks = taskList.getNumberTasks();
-        return UserInterface.formatTaskAddedResponse(response, totalTasks);
-    }
-
-    private static String handleExit() throws DukeException {
-        new Timer().schedule(new TimerTask() {
-            @Override
-            public void run() {
-                System.exit(0);
-            }
-        }, 1000);
-
-        return UserInterface.formatExit();
+    private void initController() {
+        logicController = new LogicController(taskList);
     }
 
 }
